python/requirements.txt
argparse
numpy
path.py
Pillow
pybullet
pygame
requests
tensorflow-gpu
torch
torchvision
torchviz
tqdm
